+++ static/app/components/events/autofix/preferences/hooks/useOrganizationRepositories.ts
@@ -1,12 +1,8 @@
 import {useCallback, useEffect, useMemo, useRef} from 'react';
 
-import type {Repository, RepositoryWithSettings} from 'sentry/types/integrations';
-import type {Organization} from 'sentry/types/organization';
-import {apiOptions} from 'sentry/utils/api/apiOptions';
+import type {Repository} from 'sentry/types/integrations';
 import {getApiUrl} from 'sentry/utils/api/getApiUrl';
 import {useFetchSequentialPages} from 'sentry/utils/api/useFetchSequentialPages';
-import {encodeSort} from 'sentry/utils/discover/eventView';
-import type {Sort} from 'sentry/utils/discover/fields';
 import type {ApiQueryKey} from 'sentry/utils/queryClient';
 import {useOrganization} from 'sentry/utils/useOrganization';
 
@@ -61,32 +57,5 @@ export function useOrganizationRepositories({query = {}} = {} as Props) {
       isFetching,
     }),
     [data, isFetching, rest]
-  );
-}
-
-export function organizationRepositoriesWithSettingsInfiniteOptions({
-  organization,
-  query,
-  staleTime,
-}: {
-  organization: Organization;
-  query?: {
-    cursor?: string;
-    integration_id?: string;
-    per_page?: number;
-    query?: string;
-    sort?: Sort;
-    status?: 'active' | 'deleted' | 'unmigratable';
-  };
-  staleTime?: number;
-}) {
-  const sortQuery = query?.sort ? encodeSort(query.sort) : undefined;
-  return apiOptions.asInfinite<RepositoryWithSettings[]>()(
-    '/organizations/$organizationIdOrSlug/repos/',
-    {
-      path: {organizationIdOrSlug: organization.slug},
-      query: {expand: 'settings', per_page: 100, ...query, sort: sortQuery},
-      staleTime: staleTime ?? 0,
-    }
   );
 }
